requirements.txt
django
djangorestframework
django-cors-headers
whitenoise
django-environ
djangorestframework-simplejwt
beautifulsoup4
requests
google-api-python-client
